requirements.txt
adbutils
Appium-Python-Client
streamlit
selenium
requests
python-dotenv
PyYAML
google-genai
lxml